Split up models.py into seperate files
