+++ yarn-error.log
@@ -1,0 +1,79 @@
+Arguments: 
+  /usr/bin/node /usr/local/lib/node_modules/yarn/bin/yarn.js install
+
+PATH: 
+  /home/julian/.cargo/bin:/home/julian/.local/share/umake/bin:/home/julian/bin:/home/julian/.local/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin:/usr/lib/jvm/java-8-oracle/bin:/usr/lib/jvm/java-8-oracle/db/bin:/usr/lib/jvm/java-8-oracle/jre/bin
+
+Yarn version: 
+  1.16.0
+
+Node version: 
+  8.10.0
+
+Platform: 
+  linux x64
+
+Trace: 
+  Error: EACCES: permission denied, unlink '/home/julian/Github/dienst2/dienst2/static/lib/.yarn-integrity'
+
+npm manifest: 
+  {
+    "dependencies": {
+      "bootstrap": "^3.4.1",
+      "font-awesome": "^4.7.0",
+      "jquery": "^1.12.4",
+      "select2": "^4.0.5",
+      "select2-bootstrap-theme": "^0.1.0-beta.10",
+      "themestrap": "WISVCH/bootstrap-theme#gh-pages"
+    }
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  # THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
+  # yarn lockfile v1
+  
+  
+  almond@~0.3.1:
+    version "0.3.3"
+    resolved "https://registry.yarnpkg.com/almond/-/almond-0.3.3.tgz#a0e7c95ac7624d6417b4494b1e68bff693168a20"
+    integrity sha1-oOfJWsdiTWQXtElLHmi/9pMWiiA=
+  
+  bootstrap@^3.4.1:
+    version "3.4.1"
+    resolved "https://registry.yarnpkg.com/bootstrap/-/bootstrap-3.4.1.tgz#c3a347d419e289ad11f4033e3c4132b87c081d72"
+    integrity sha512-yN5oZVmRCwe5aKwzRj6736nSmKDX7pLYwsXiCj/EYmo16hODaBiT4En5btW/jhBF/seV+XMx3aYwukYC3A49DA==
+  
+  font-awesome@^4.7.0:
+    version "4.7.0"
+    resolved "https://registry.yarnpkg.com/font-awesome/-/font-awesome-4.7.0.tgz#8fa8cf0411a1a31afd07b06d2902bb9fc815a133"
+    integrity sha1-j6jPBBGhoxr9B7BtKQK7n8gVoTM=
+  
+  jquery-mousewheel@~3.1.13:
+    version "3.1.13"
+    resolved "https://registry.yarnpkg.com/jquery-mousewheel/-/jquery-mousewheel-3.1.13.tgz#06f0335f16e353a695e7206bf50503cb523a6ee5"
+    integrity sha1-BvAzXxbjU6aV5yBr9QUDy1I6buU=
+  
+  jquery@^1.12.4:
+    version "1.12.4"
+    resolved "https://registry.yarnpkg.com/jquery/-/jquery-1.12.4.tgz#01e1dfba290fe73deba77ceeacb0f9ba2fec9e0c"
+    integrity sha1-AeHfuikP5z3rp3zurLD5ui/sngw=
+  
+  select2-bootstrap-theme@^0.1.0-beta.10:
+    version "0.1.0-beta.10"
+    resolved "https://registry.yarnpkg.com/select2-bootstrap-theme/-/select2-bootstrap-theme-0.1.0-beta.10.tgz#b9426ecfc03bf4a235e76a132377574310469ac0"
+    integrity sha1-uUJuz8A79KI152oTI3dXQxBGmsA=
+  
+  select2@^4.0.5:
+    version "4.0.5"
+    resolved "https://registry.yarnpkg.com/select2/-/select2-4.0.5.tgz#7aac50692561985b34d3b82ec55e226f8960d40a"
+    integrity sha1-eqxQaSVhmFs007guxV4ib4lg1Ao=
+    dependencies:
+      almond "~0.3.1"
+      jquery-mousewheel "~3.1.13"
+  
+  themestrap@WISVCH/bootstrap-theme#gh-pages:
+    version "0.3.0"
+    resolved "https://codeload.github.com/WISVCH/bootstrap-theme/tar.gz/21cc92a37f3efeae03c2e4464260d61033c0deb1"
